No longer requires Odin
Changed it to use lists instead of dictionaries (which would not serialize at edit time and get cleared on play causing no equipment to work). Works great now.

Uses NaughtInspectorAttributes for a precache button, which is free. Comment that out if you don't wanna use it and just want to use the right click context menu to precache everything